feat: integrate FdmTaskDialog component and streamline task dialog logic
- Introduced FdmTaskDialog component to encapsulate task creation logic.
- Removed redundant task dialog configuration from FDM.vue.
- Updated task handling methods to utilize the new dialog component.
- Enhanced localization strings for task creation feedback.
- Improved reactivity and state management for task-related data.
