+++ packages/ldp/src/locale/lang/en.js
@@ -5,7 +5,8 @@ export default {
   packages_ldp_view_lineage: 'View Lineage',
   packages_ldp_lineage_loading_tips: 'Double-click the node to drill down.',
   packages_ldp_table_comment: 'Table Comment',
-  packages_ldp_src_dashboard_anEsctui: 'Press Esc to exit the traceability scene',
+  packages_ldp_src_dashboard_anEsctui:
+    'Press Esc to exit the traceability scene',
   packages_ldp_src_tablepreview_querenshanchu: 'Confirm to delete? ',
   packages_ldp_src_tablepreview_gaibiaojianghuicong:
     'The table will be deleted from the database and cannot be recovered after the operation',
@@ -18,9 +19,11 @@ export default {
     'A data hub allows you to consolidate mission critical data from siloed sources into centralized storage, then provide fresh data to downstream applications or dashboards from a single location.',
   packages_ldp_data_hub_intro_desc2:
     'Tapdata uses CDC technology to sync the data from source and uses MongoDB / MongoDB Atlas as hub storage, to achieve the near real time data latency experiecne.',
-  packages_ldp_data_hub_intro_scene_title: 'Which Use Cases Can I Use Real Time Data Hub?',
+  packages_ldp_data_hub_intro_scene_title:
+    'Which Use Cases Can I Use Real Time Data Hub?',
   packages_ldp_data_hub_intro_scene_single_view: 'Single View',
-  packages_ldp_data_hub_intro_scene_single_view_sub: 'Products ｜ Customers ｜ Orders',
+  packages_ldp_data_hub_intro_scene_single_view_sub:
+    'Products ｜ Customers ｜ Orders',
   packages_ldp_data_hub_intro_scene_realtime: 'Real Time',
   packages_ldp_data_hub_intro_scene_realtime_sub: 'Dashboards  ｜  Reports',
   packages_ldp_data_hub_intro_scene_api: 'Enterprise API Service',
@@ -39,8 +42,10 @@ export default {
   page_title_data_hub: 'Real Time Data Hub',
   packages_ldp_source_empty_text: '1. Create your own data source first',
   packages_ldp_target_empty_text: '2. Next, create your target database.',
-  packages_ldp_not_support_increments: 'Current source data does not support increments.',
-  packages_ldp_drag_source_table_to_start: 'Drag the source table here to start copying.',
+  packages_ldp_not_support_increments:
+    'Current source data does not support increments.',
+  packages_ldp_drag_source_table_to_start:
+    'Drag the source table here to start copying.',
   packages_ldp_run_only_once: 'Run only once',
   packages_ldp_run_every_10_minutes: 'Run every 10 minutes',
   packages_ldp_run_every_hour: 'Run every 1 hour',
@@ -50,5 +55,9 @@ export default {
   packages_ldp_book_demo: 'Feel free to Book a Demo with us',
   packages_ldp_mdm_create_method: 'Create Method',
   packages_ldp_mdm_create_method_transformation: 'Use Data Transformation Task',
-  packages_ldp_mdm_create_method_materialized: 'Use Materialized View'
+  packages_ldp_mdm_create_method_materialized: 'Use Materialized View',
+  packages_ldp_fdm_create_task_has_synced:
+    'You have already copied some tables, but you can still select more. Newly selected tables will be added to the existing list.',
+  packages_ldp_fdm_create_task_has_expression:
+    'You have selected tables by regex matching and cannot select more. Please use the Data Copy feature to continue copying tables.',
 }
